requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
databases==0.9.0
alembic==1.14.0
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic[email]==2.10.4
pydantic-settings==2.7.0
postgres==4.0
passlib[bcrypt]==1.7.4
sqlalchemy-utils==0.41.2
pyjwt==2.10.1
httpx==0.28.1
testcontainers==4.9.0
deepgram-sdk
langchain
langchain-openai
langchain-community
